Made minor changes but error needs to be resolved
